+++ CHANGELOG.md
@@ -5,6 +5,13 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.4] - 2023-04-27
+
+### Fixed
+
+- Select fields had some timing issue, but should now work properly with client-side validation.
+- Fixed a few problems with error+tainted checking for client-side validation.
+
 ## [0.8.3] - 2023-04-26
 
 ### Fixed